VTCCA-631 update assessments endpoint to return all the property link data
